Realise authentication and authorization

+++ web/src/main/kotlin/commentaries/controllers/CommentaryController.kt
@@ -4,10 +4,15 @@ import commentaries.extensions.toResponseCommentary
 import commentaries.models.CommentaryCreationRequest
 import commentaries.models.CommentaryResponse
 import commentaries.services.ICommentaryService
+import exceptions.NotEnoughRightsException
+import org.eclipse.microprofile.jwt.Claim
+import org.eclipse.microprofile.jwt.Claims
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponses
 import java.util.*
+import javax.annotation.security.RolesAllowed
 import javax.enterprise.context.RequestScoped
+import javax.inject.Inject
 import javax.ws.rs.DELETE
 import javax.ws.rs.GET
 import javax.ws.rs.POST
@@ -19,12 +24,21 @@ class CommentaryController(
     private var _commentaryService: ICommentaryService,
 ) {
 
+    @Inject
+    @Claim(standard = Claims.upn)
+    private lateinit var userId: String
+
+    @Inject
+    @Claim(standard = Claims.groups)
+    private lateinit var groups: Set<String>
+
     @APIResponses(
         APIResponse(responseCode = "200", description = "Commentary is created"),
         APIResponse(responseCode = "400", description = "Invalid title or body"),
         APIResponse(responseCode = "404", description = "User with sent id does not exist")
     )
     @POST
+    @RolesAllowed("User", "HR", "Admin")
     fun add(commentaryCreationRequest: CommentaryCreationRequest): CommentaryResponse {
         val creationModel = commentaryCreationRequest.toCreationModel()
 
@@ -37,6 +51,7 @@ class CommentaryController(
     )
     @GET
     @Path("/{id}")
+    @RolesAllowed("Admin")
     fun getById(id: UUID): CommentaryResponse {
         val commentary = _commentaryService.getById(id)
 
@@ -48,6 +63,7 @@ class CommentaryController(
     )
     @GET
     @Path("/user/{userId}")
+    @RolesAllowed("Admin")
     fun getAllByUserId(userId: UUID): List<CommentaryResponse> {
         val commentaries = _commentaryService.getAllByUserId(userId)
 
@@ -60,7 +76,14 @@ class CommentaryController(
     )
     @DELETE
     @Path("/{id}")
+    @RolesAllowed("User", "HR", "Admin")
     fun removeById(id: UUID): CommentaryResponse {
+        val commentary = _commentaryService.getById(id)
+
+        if (!groups.contains("Admin") && commentary.userId != UUID.fromString(userId)) {
+            throw NotEnoughRightsException("You do not have access to this commentary!")
+        }
+
         return _commentaryService.removeById(id).toResponseCommentary()
     }
 }
